+++ spec/support/has_an_alias.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+# :nocov:
 RSpec.shared_examples "has an alias" do |aliased_name, original_name|
   describe "##{aliased_name}" do
     it "is an alias of ##{original_name}" do
@@ -7,3 +8,4 @@ RSpec.shared_examples "has an alias" do |aliased_name, original_name|
     end
   end
 end
+# :nocov: