+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 **Changed**:
 - Rewrite `BruteForce` calculator to be an order of magnitude faster. Uses `Enumerator::Product` if available, otherwise a better implementation than before.
 
+**Fixed**:
+- Allow calling calculators with an empty list, returning an empty hash.
+
 [Compare v0.13.1...main](https://github.com/trinistr/dicey/compare/v0.13.1...main)
 
 ## [v0.13.1] — 2025-09-03